requirements.txt
Flask
Gunicorn
flask_pymongo
flask_mail
pillow
pymongo
Crypto
pycryptodome
numpy
